+++ src/views/AuthorityView/AuthorityView.js
@@ -11,8 +11,10 @@ import {
   useIntl,
   FormattedMessage,
 } from 'react-intl';
+import queryString from 'query-string';
 import cloneDeep from 'lodash/cloneDeep';
 import set from 'lodash/set';
+import omit from 'lodash/omit';
 
 import {
   LoadingView,
@@ -59,9 +61,13 @@ const AuthorityView = ({
     () => {
       setSelectedAuthorityRecordContext(null);
 
+      const parsedSearchParams = queryString.parse(location.search);
+      const commonSearchParams = omit(parsedSearchParams, ['authRefType', 'headingRef']);
+      const newSearchParamsString = queryString.stringify(commonSearchParams);
+
       history.push({
         pathname: '/marc-authorities',
-        search: location.search,
+        search: newSearchParamsString,
       });
     },
     // eslint-disable-next-line react-hooks/exhaustive-deps